cli: make channels specific utilities local to route
