--- raiden-cli/src/utils/channels.ts
@@ -1,44 +1,0 @@
-import { RaidenChannel, RaidenChannels, ChannelState } from 'raiden-ts';
-import { ApiChannel, ApiChannelState } from '../types';
-
-export function flattenChannelDictionary(channelDict: RaidenChannels): RaidenChannel[] {
-  // To flatten structure {token: {partner: [channel..], partner:...}, token...}
-  return Object.values(channelDict).reduce(
-    (allChannels, tokenPartners) => allChannels.concat(Object.values(tokenPartners)),
-    [] as RaidenChannel[],
-  );
-}
-
-function transformChannelStateForApi(state: ChannelState): ApiChannelState {
-  let apiState;
-  switch (state) {
-    case ChannelState.open:
-    case ChannelState.closing:
-      apiState = ApiChannelState.opened;
-      break;
-    case ChannelState.closed:
-    case ChannelState.settleable:
-    case ChannelState.settling:
-      apiState = ApiChannelState.closed;
-      break;
-    case ChannelState.settled:
-      apiState = ApiChannelState.settled;
-      break;
-  }
-  return apiState;
-}
-
-export function transformChannelFormatForApi(channel: RaidenChannel): ApiChannel {
-  return {
-    channel_identifier: channel.id.toString(),
-    token_network_address: channel.tokenNetwork,
-    partner_address: channel.partner,
-    token_address: channel.token,
-    balance: channel.capacity.toString(),
-    total_deposit: channel.ownDeposit.toString(),
-    total_withdraw: channel.ownWithdraw.toString(),
-    state: transformChannelStateForApi(channel.state),
-    settle_timeout: channel.settleTimeout.toString(),
-    reveal_timeout: '50', // FIXME: Not defined here. Python client handles reveal timeout differently,
-  };
-}
